Updated to npmjs.com
https://www.npmjs.com/package/ogf-nsi-collector

It seems that @ogflab, modified from @ideditor/nsi-collector, needs to create an organization. However, I rarely use the npm registry, so I don't know how to create one or what it is used for other than this package. Therefore, I decided to remove it.

I don't know if there will be a chance to create @ogflab in the future, nor do I know if this package can be renamed. (It's not possible on PyPi; only deletion is allowed and versions cannot collide because PyPi mirrors will always keep them.)